Cleaned up the structure for the api call by creating separate files
